remove debugger option from capp project and added the debug logic to the esb debug profile
