+++ plugins/org.wso2.developerstudio.eclipse.carbonserver44microei/META-INF/MANIFEST.MF
@@ -66,4 +66,5 @@ Import-Package: javax.xml.stream;version="1.0.1.wso2v1",
  org.wso2.developerstudio.eclipse.utils.file;version="[4.1.0,7.0.0)",
  org.wso2.developerstudio.eclipse.utils.wst;version="[4.1.0,7.0.0)"
 Bundle-Vendor: WSO2
-Export-Package: org.wso2.developerstudio.eclipse.carbonserver44microei.register.product.servers;version="6.5.0"
+Export-Package: org.wso2.developerstudio.eclipse.carbonserver44microei.register.product.servers;version="6.5.0",
+ org.wso2.developerstudio.eclipse.carbonserver44microei.wizard
