--- plugins/org.wso2.developerstudio.eclipse.carbonserver44microei/src/org/wso2/developerstudio/eclipse/carbonserver44microei/wizard/ExportDistributionAndDebugAction.java
@@ -1,50 +1,0 @@
-/*
- * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
-
- *      http://www.apache.org/licenses/LICENSE-2.0
-
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.developerstudio.eclipse.carbonserver44microei.wizard;
-
-import org.eclipse.jface.action.IAction;
-import org.eclipse.jface.viewers.ISelection;
-import org.eclipse.jface.viewers.IStructuredSelection;
-import org.eclipse.jface.wizard.WizardDialog;
-import org.eclipse.ui.IActionDelegate;
-import org.eclipse.ui.PlatformUI;
-
-
-public class ExportDistributionAndDebugAction implements IActionDelegate {
-
-	IStructuredSelection selection;
-
-	public void run(IAction action) {
-		if (selection != null) {
-			DistributionProjectExportAndDebugWizard wizard = new DistributionProjectExportAndDebugWizard();
-			wizard.init(PlatformUI.getWorkbench(), selection);
-			WizardDialog exportWizardDialog = new WizardDialog(
-					PlatformUI.getWorkbench().getActiveWorkbenchWindow().getShell(), wizard);
-			exportWizardDialog.open();
-		}
-	}
-
-	public void selectionChanged(IAction action, ISelection selection) {
-		if (selection instanceof IStructuredSelection) {
-			this.selection = (IStructuredSelection) selection;
-		}
-
-	}
-
-}
